requirements.txt
opencv-python
kafka
flask
scikit-image
tensorflow
config
Pillow